add rule for percentliteral delimeters
I grepped the source code looking for actual usage and went through
the old ruby 1.9 pickaxe PDF and the result is that '{}' is how we
actually have written the chef source code and its what all the old
farts actually learned.

e.g.

% egrep -rI "%w" . | sed -e 's/.*%w/%w/' | cut -c 1-3 | sort | uniq -c | sort -rn
 164 %w{
 100 %w(
  54 %w[
   1 %w|

I vote the old ways are best and there's no reason to change this.