Enable Style/Encoding cop
The encoding cop was previously disabled as it *added* the encoding comments. Now it removes them if they were setting utf-8 since that's the default on Ruby 2 and is totally pointless to set.

This also enables the frozen string literal comment cop for this repo so
we can start the process of getting ready to enable that.

Signed-off-by: Tim Smith <tsmith@chef.io>