Comment out obsolete/renamed lints
In the course of updating the pinned RuboCop versions, there hadn't been
changes to the enabled upstream defaults in chefstyle.yml, and so they've been
silently broken. This becomes noticeable when running `rubocop -c
config/chefstyle.yml`.

The decisions on how I did this:
- Enabled cops renamed? Comment out for now to keep parity, the re-enabling of those cops should be done in a
separate PR to this one (which is designed to simplify the chefstyle
internals).
- Disabled cops renamed? Update name, keep parity.
- Cop removed entirely? Remove cop from config, keep parity.

This is all to say that this commit changes nothing:
```
$ ./bin/chefstyle --show-cops | sha256sum
faa8fd05273862f47447a5357b8001976e8badb8d53c4e761f046383396930b5  -
```

Signed-off-by: David Crosby <dcrosby@fb.com>