[Chefstyle] Mark Chefstyle repository as deprecated
Chefstyle was merged into Cookstyle in
https://github.com/chef/cookstyle/pull/968, and so there's no longer a
reason to keep a separate repository. Any future fixes to the Chefstyle
configuration should be handled in the Cookstyle repository.

Once all the Chef github org repositories and major reverse dependencies on the
gem (https://rubygems.org/gems/chefstyle/reverse_dependencies) are addressed,
it can be sent to chef-boneyard.

Signed-off-by: David Crosby <dcrosby@fb.com>